Move remote API classes to `QualityGateResult`.
