+++ src/main/java/io/jenkins/plugins/util/QualityGateResult.java
@@ -6,7 +6,10 @@ import java.util.Collection;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import org.kohsuke.stapler.export.Exported;
+import org.kohsuke.stapler.export.ExportedBean;
 import org.jenkinsci.plugins.scriptsecurity.sandbox.whitelists.Whitelisted;
+import hudson.model.Result;
 
 /**
  * Result of a quality gate evaluation. Aggregates the individual results of the quality gates into an overall status.
@@ -99,6 +102,75 @@ public class QualityGateResult implements Serializable {
 
         public String getActualValue() {
             return actualValue;
+        }
+    }
+
+    /**
+     * Remote API to list the overview of the quality gate evaluation.
+     */
+    @ExportedBean
+    public static class QualityGateResultApi {
+        private final QualityGateResult qualityGateResult;
+
+        /**
+         * Creates a new instance of {@link QualityGateResultApi}.
+         *
+         * @param qualityGateResult
+         *         the quality gate result to show
+         */
+        public QualityGateResultApi(final QualityGateResult qualityGateResult) {
+            this.qualityGateResult = qualityGateResult;
+        }
+
+        @Exported(inline = true)
+        public QualityGateStatus getOverallResult() {
+            return qualityGateResult.getOverallStatus();
+        }
+
+        @Exported(inline = true)
+        public Collection<QualityGateItemApi> getResultItems() {
+            return qualityGateResult.getResultItems()
+                    .stream()
+                    .map(QualityGateItemApi::new)
+                    .collect(Collectors.toList());
+        }
+    }
+
+    /**
+     * Remote API to show the content of an individual quality gate item.
+     */
+    @ExportedBean
+    public static class QualityGateItemApi {
+        private final QualityGateResultItem item;
+
+        /**
+         * Creates a new instance of {@link QualityGateItemApi}.
+         *
+         * @param item
+         *         the quality gate result item to show
+         */
+        public QualityGateItemApi(final QualityGateResultItem item) {
+            this.item = item;
+        }
+
+        @Exported
+        public String getQualityGate() {
+            return item.getQualityGate().getName();
+        }
+
+        @Exported
+        public double getThreshold() {
+            return item.getQualityGate().getThreshold();
+        }
+
+        @Exported(inline = true)
+        public Result getResult() {
+            return item.getStatus().getResult();
+        }
+
+        @Exported
+        public String getValue() {
+            return item.getActualValue();
         }
     }
 }
